Intelligent-SHL-Assessment-Recommendation-System-main/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
streamlit==1.28.1
requests==2.31.0
python-dotenv==1.0.0
groq>=0.11.0
sentence-transformers>=2.3.0
faiss-cpu==1.7.4
pandas==2.1.3
beautifulsoup4==4.12.2
lxml==4.9.3
pydantic>=2.5.0,<3.0.0
pydantic-core>=2.14.0,<3.0.0
openpyxl==3.1.2
tqdm==4.66.1

